numba jit for kernel functions, refactor, + dtype cast for dtype consistency

+++ src/grnn/layers.py
@@ -9,5 +9,5 @@ class SummationLayer:
     def transform(self, K):
         denom = np.sum(K, axis=1)
         nom = np.matmul(K, self.y_)
-        out = np.divide(nom, denom, out=np.zeros_like(nom, dtype=np.float64), where=denom > 0)
+        out = np.divide(nom, denom, out=np.zeros_like(nom, dtype=K.dtype), where=denom > 0)
         return out
